refactor transfer encryptions into separate submodule
